refactor(tui): extract app tests to app/tests.rs (S4 PR 3)
Third slice. The 1572-line `#[cfg(test)] mod tests { ... }` block at the
bottom of `app/mod.rs` moves to its own file. No behaviour change;
84/84 TUI tests pass.

Mechanics:
  - new file: crates/memphis-tui/src/app/tests.rs (~1576 LoC)
  - app/mod.rs: 4100 → 2530 LoC (-1570)
  - mod.rs end now reads:
      #[cfg(test)]
      mod tests;
  - tests.rs preserves the same `use super::{...}` imports the inline
    block had — child module sees parent's private items so nothing
    needed to become pub(super)
  - 4-space indentation peeled off so file looks like a normal Rust
    test file rather than an accidentally-nested-mod

Why this slice:
  - Pure test extraction is the lowest-risk diff per LoC moved
  - 1500 LoC is the ADR-001 ceiling for src/; while crates/ allows up
    to 2000, mod.rs at 4100 was over even that. Now down to 2530;
    one more slice of production code (worker/screen/render) is
    needed to clear the 2000 ceiling.

Status after S4 PRs 1-3 on this branch:
  app/mod.rs          2530 LoC  (target <2000 for crates/)
  app/host_results.rs 1323 LoC
  app/tests.rs        1576 LoC
  app/commands.rs      671 LoC
